requirements.txt
nump>==1.13.3
scipy>=0.19.1
nibabel>=2.1.0
pytest-cov>=2.5.1
pillow>=4.3.0
cython>=0.27.1
tensorflow>=1.3.0
